news-aggregator/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
httpx==0.28.1
feedparser==6.0.11
python-dotenv==1.1.1
